compose: Parameterize MCP server port more thoroughly
The mcp server port is specified 4 different times
(3 in compose.yaml and once in the mcp-atlassian.env template).

This commit consolidates it to one place: the compose.env template,
so it gets defined in the .env file.

At the same time that this commit drops the PORT definition from
mcp-atlassian.env, it also drops the TRANSPORT definition because
the transport is already explicitly given on the command line.