recipes: Add test-package recipe
Now that we have an MCP server for interacting with
testing farm, we need a recipe for actually testing
packages with it.

This commit adds a single bare bones recipe for starting a
test for a specific package against a specific compose and
waiting for the results.

A future commit will add more sophisticated recipes for testing
chains of packages based on reverse dependencies.