Makefile: Move defaults to top of file
At the moment, the defaults are grouped with the recipe definitions,
which on the surface makes sense, because it's putting logically
related things together.

Doing things this way, however, has two downsides:

  1. A casual reader has to scroll deep into the makefile to find the
     defaults
  2. If defaults from competing recipes have common variable names,
     the "first one wins" for specifying the default.

That latter issue is particularly problematic because it means that
different recipes can't specify a different default package without
namespacing the package variable name, which would harm readability.

This commit addresses both problems by moving default definitions to
the top of the file and properly marking them per-recipe or global as
appropriate.

Since parts of the defined targets are being reorganized to the top of
the file, this commit also regroups the phony target specifications
there as well, so the available targets are documented in the preamble
of the Makefile.