Makefile: Use `podman exec` to start goose
Right now the readme recommends starting the
Atlassian MCP server manually before running
the goose REPL or goose container shell, so that
the MCP services don't get tied to the interactive
console and get in the way.

That goes against the spirit of the dependencies
specified in compose.yaml, however, and creates an
extra step that can be somewhat cumbersome. Furthermore,
the problem is only going to get exacerbated in the
future as we add new MCP servers.

This commit updates the Makefile to use `podman up -d`
and `podman exec` instead of `podman run` so all the
containers get started in the background, and only
the shell itself is connected to the active terminal.