templates: Provide .env template as well
commit 9cee6aefabc99556960fee28af8313e4a08cfc84 split the various
bits of configuration for each service into separate .env files,
so the secrets for e.g. jira don't leak into the container for
goose.

Clearly, this is better for security, but it means there is now
no template for the toplevel .env file. The .env file will be
important because it provides a means to provide configuration
about the services specific to how they're invoked from
compose.yaml.

This commit adds an empty compose.env to templates/ to serve
the purpose of being a template for .env. The commit updates
the Makefile to instantiate this new template in the same
rule it instantiates the other templates, and renames that rule
from "secrets" to "config" to reflect its no longer acting
on just templates that harbor secrets.