container: Add way to build goose from source
Goose was exiting unexpectedly sometimes with either no error message or
a vague one. To find out what was going on I needed to build goose from
source with a patch applied to instrument the code.

The issue ultimately ended up being a 600 second timeout hard coded in
the source, but finding that out required rearchitecting a bit how the goose
container was built.

This commit (and the commit directly before this one) are an evolution of that
effort. It adds a new stage to the containerfile for building goose from
source, and a stage for hosting the built goose. *.patch from the
goose-container/ folder are applied to the checked out goose tree.

This commit also adds a new Makefile target `debug-build` for producing
the output image to fire things off.