Temporarily pin beeai to avoid patch application issues
With the new version of beeai being released, our patch fails to apply, making building new images impossible.
As we are waiting for the patch being merged upstream, temporarily pinning until
that code is merged and released should be enough. Alternatively, we could regenerate the patch for each new release touching the code.