reverse-dependences: Fix bug in find-package-dependents script
find-package-dependents was passing the user's max_results limit
to the generate_direct_dependents function. This would be fine,
except we sometimes need more than max_results generated if some
of the results are filtered by the user's filter command.

The upshot is the script would sometimes traverse the dependents
tree deep instead of wide, leading to performance issues.

This commit drops the max_result limit on generate_direct_dependents,
since it's wrong and not needed anyway (because max_results is
already handled elsewhere).

This commit also adds a couple of test cases to make sure the problem
doesn't creep in again.