Enable parallel tool calls
The triage agent, at the beginning of its execution, tends to run two tools
in parallel - `think` and `get_jira_details`. That's disallowed by
default, but I don't really see a reason why not to allow it. If it
proves problematic, we can disable it and try to overcome the problem
by using requirements.

Signed-off-by: Nikola Forró <nforro@redhat.com>