supervisor/main.py: Initialize kerberos ticket before use
Use the init_kerberos_ticket() function to ensure a valid Kerberos ticket
is available before making any erratatool API calls.

This change makes the KEYTAB_FILE environment variable work correctly for
the supervisor - if it's set, then the supervisor will use it to
authenticate to Kerberos. Without the envvar it produces a clean
error rather than failing with a mysterious stack trace when talking
to erratatool.