Reorganize secrets and keymaps
Instead of primarily organizing around services, use secrets and
configmaps that are organized by topic - e.g. chat, jira, kerberos;
so environment variables related to model access are now in
configmap/chat-env, and kerberos-related environment variables are in
configmap/kerberos-env.

URLs to known services within the cluster are put into the
endpoints-env configmap; e.g, MCP_GATEWAY_URL, REDIS_URL, etc.

We now mostly import these variables using envFrom in the
various deployment and cronjob definitions.

To allow consistent paths to be used in the containers, we
standardize on /etc/secrets for secrets - so e.g.
/etc/secrets/jotnar-vertex-prod.json instead of
/home/beeai/jotnar-vertex-prod.json

The SSE_PORT value from the old configmap-x-mcp-gateway is
now just folded into the deployment config, since it was
not really configuration that could be meaningfully changed.